requirements.txt
# 核心
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
pytorch-lightning>=2.3.0
torchmetrics>=1.4.0

# 配置与工具
omegaconf>=2.3.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0.1
rich>=13.7.0

# 导出
onnx>=1.15.0
onnxruntime>=1.16.0
